docs/requirements.txt
# Web Framework
fastapi==0.135.3
uvicorn[standard]==0.44.0

# Database
sqlalchemy==2.0.49
psycopg2-binary==2.9.10

# Authentication & Security
pyjwt==2.10.1
argon2-cffi==23.1.0

# Validation
pydantic[email]==2.12.5
pydantic-core==2.41.5
google-genai==1.72.0


